+++ src/main/java/com/app/rdc/taxation/item/objects/Medical.java
@@ -7,10 +7,10 @@ import lombok.Setter;
 @Setter
 public class Medical implements Item {
     private String name;
-    private double price;
+    private float price;
     private boolean imported;
 
-    public double getTax() {
+    public float getTax() {
 
         if(imported == false)
             return 0f;
